modal_app/requirements-light.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
lxml>=5.0.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0
fastapi>=0.115.0
pydantic>=2.0.0
jsonschema>=4.0.0
